+++ src/main/java/com/multi/vo/PmediVo.java
@@ -48,6 +48,6 @@ public class PmediVo {
 		this.id = id;
 		this.name = name;
 	}
-	
+
 	
 }
